+++ LYToolsKit/Foundation/NSString/NSString+LYURL.h
@@ -24,7 +24,6 @@ extern NSString *const kLYURLFragment;
 /**
  解析URL
  格式如：http://host.name:8888/test/page/?keyA=valueA&keyB=valueB#top
- URL中的 参数表
  @param    kLYURLScheme   如 http
  @param    kLYURLHost     如 host.name
  @param    kLYURLPort     如 8888
@@ -39,20 +38,29 @@ extern NSString *const kLYURLFragment;
 
 /**
  URL地址Reserved Characters与Unreserved Characters之外的字符做百分号编码。
- 编码范围：
+ @code编码：
  1. UTF-8编码
- 2. !*'();:@&=+$,/?%#[ ]除外字符
+ 2. !*'();:@&=+$,/?%#[ ]
  */
 - (instancetype)ly_stringAllAddingPercentEncoding;
-/** 处理字符串中的中文、空格、《》（）【】 */
+/**
+ 处理特殊字符
+ @code
+ 1. 编码中文、《》、（）、【】
+ 2. 空格替换为空字符
+ */
 - (instancetype)ly_stringAddingPercentEscapesChineseSpace;
 /**
  根据 string构建 URL实例
 
- @param string string URL String
- @return return value URL实例
+ @param string URL String
+ @return return URL实例
  */
 + (NSURL *)ly_URLWithString:(NSString *)string;
+/** url拼接参数，跳过已经存在的参数值 */
++ (nullable instancetype)ly_stringURL:(NSString *)url appendNotExistParams:(nullable NSDictionary<NSString *, NSString *> *)params;
+/** url拼接参数，覆盖已经存在的参数值 */
++ (nullable instancetype)ly_stringURL:(NSString *)url appendParams:(nullable NSDictionary<NSString *, NSString *> *)params;
 
 @end