+++ CHANGELOG.md
@@ -6,3 +6,6 @@
 -【ADD】屏幕尺寸宏
 ###【0.1.1】
 -【FIX】iOS 12.1刘海屏判断‘ly_hasFringeScreen’错误问题解决
+### [0.2.0]
+-【MDF】'NSString+LYURL'方法描述修改
+-【ADD】URL拼接参数功能‘stringURL:appendNotExistParams:’、‘stringURL:appendParams:’